nullable

+++ groovy-impl/src/main/java/org/jetbrains/plugins/groovy/lang/completion/GroovyCompletionUtil.java
@@ -524,7 +524,7 @@ public class GroovyCompletionUtil {
       newStartOffset = marker.getStartOffset();
     }
 
-    if (toDelete.isValid()) {
+    if (toDelete != null && toDelete.isValid()) {
       document.deleteString(toDelete.getStartOffset(), toDelete.getEndOffset());
     }
